Adjust route regex & parameter bindings
This will allow both the width and height to pass route verification as negative values. Again this is just to catch the error better and provide a more verbose/accurate error message. With this idea in mind, it will also accept 5 character sizes
I’m also adding int as parameter bindings to ensure they are always a numeric value.