Build a Subdomain model
This will be an App model and not a database model. It's simply a way to contain the business logic of a single subdomain entity.

Utilizing this will fundamentally change the Apps current execution architecture. The FillSet will need to be refactored into FillSettings and attached to their respective subdomain. Then instead of the app having a global Collection of FillSets it will have a Subdomain Service providing the active subdomains and they each providing their own fill settings.